Hoist shared host resolution helpers into host/core.ts
Move the genuinely-shared resolution layer out of host.ts into host/core.ts:
the auth-lens/handle resolvers (lensDb, lensRtdb, sessionDb, sessionRtdb,
ensureRtdb, authStateForLens, lensCacheKey, lensProvenance), descriptor
resolution (resolveTarget, resolveConstraint), snapshot serialization
(serializeDocSnap), and opProvenance. host.ts now imports them. Bodies moved
verbatim; the orphaned pyric/firestore query-builder and database imports are
dropped from host.ts. No behavior change.

The pre-existing unused Sandbox / SandboxError imports in host.ts are left
untouched (out of scope).
